2008-06-06  Andreas Matthies  <andreas.matthies@gmx.net>
	* raster/raster-sprite-status.c, raster-sprite-status.c,
	raster-sprite.c, raster-sprite.h, vicii/vicii-sprites.c,
	vicii-sprites.h, vicii.c: Fixed warnings. Fixed a bug in sprite
	initialization using different border mode.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@18842 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/fullscrn.c
@@ -586,7 +586,7 @@ BOOL CALLBACK dialog_fullscreen_proc(HWND hwnd, UINT msg, WPARAM wparam,
 
     switch (msg) {
       case WM_NOTIFY:
-        if (((NMHDR FAR *)lparam)->code == PSN_APPLY) {
+        if (((NMHDR FAR *)lparam)->code == (UINT)PSN_APPLY) {
             fullscreen_dialog_end();
             SetWindowLong(hwnd, DWL_MSGRESULT, FALSE);
             return TRUE;
